backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.11
python-dotenv==1.0.1
alembic==1.13.1
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
